First version of react-flask one-click building
First version of react-flask one-click building, but it still has bugs.

+++ WebUI/React_version/find-my-prof/app.py
@@ -5,31 +5,34 @@ from pathlib import Path
 import shutil
 
 
+# delete the old react build and build a new one
+# this is a bit of a hack, but it works
 def build_react():
+    # create folders for react's templates and static
     tp = Path(__file__).parent/'templates'
-    if tp.exists():
-        shutil.rmtree(str(tp))
-    tp.mkdir()
-    
     st = Path(__file__).parent/'static'
-    if st.exists():
-        shutil.rmtree(str(st))
-        
-    # just in case
-    build = Path(__file__).parent/'build'
-    if build.exists():
-        shutil.rmtree(str(build))
+    # refresh the old react's templates and static
+    if tp.exists(): shutil.rmtree(str(tp))
+    tp.mkdir()
+    if st.exists(): shutil.rmtree(str(st))
     
+    # just in case there is a build folder, delete it
+    build = Path(__file__).parent/'build'
+    if build.exists(): shutil.rmtree(str(build))
+    
+    # build the react app
     os.chdir(str(Path(__file__).parent/'src'))
     os.system("npm run build")
     
+    # move the build to the correct location: template to templates, static to static
+    # static no need build, simply move it
     react_st = build/'static'
     shutil.move(str(react_st), str(Path(__file__).parent))
-    
     for i in build.iterdir():
         shutil.move(str(i), str(Path(__file__).parent/'templates'))
     
     os.rmdir(str(build))
+    
     return
 
 
@@ -44,7 +47,7 @@ def Greet(username):
 
 @app.route("/search", methods=['GET'])
 def search():
-    
+
     return render_template('search.html', page_name="FindYourProf")
 
 @app.route("/search/submit", methods=['GET', 'POST'])
@@ -62,7 +65,7 @@ def home_page():
 
 
 if __name__ == '__main__':
-    if os.environ.get('WERKZEUG_RUN_MAIN') != 'true':
-        build_react()
-        
+    # if os.environ.get('WERKZEUG_RUN_MAIN') != 'true':
+    #     build_react()
+    
     app.run(port = 8080, debug=True)
